Rename namespace JsTestAdapter.TestServer to JsTestAdapter.TestServerClient
